Now only specifically works with B/K timestreams with a single pair including detector pointing
Previously it would take in an entire B/K tod file as input, but that file is too large to handle in GitHub.  Now it takes a smaller file as input which also includes the detector pointing, which avoids having to do the complicated pointing calculation in our code.
Also included real sample data.